fix: hold the call to the keys it asked under, and to whole milliseconds
The answers are sorted by walking the asked keys, so a question the
endpoint skipped is named rather than vanishing, and the map they land
in has no prototype for a key called __proto__ to replace. The keys
themselves are read through their descriptors and pass through the
same treatment a search gives its query before reaching a sentence the
model re-reads every turn.

The deadline is truncated, because AbortSignal.timeout refuses a
fraction that the config schema accepted -- which now also declares the
integer. A 422 is the one status the shared table reads as ours: this
request is the model's, so it may write another.

The smoke spec measures the state against the prompt it answers and
the shortest option rather than the longest.

+++ packages/core/src/config/loader.ts
@@ -266,11 +266,12 @@ const agentConfigSchema = z.object({
   /**
    * How long ONE DELIVERY of the judgement call may take, in milliseconds.
    *
-   * The call declines replay, so this bounds the wait rather than a third of
-   * it. Measured 2026-09-23: 285-395 ms once connected, 1181 ms on the first
-   * call of a process.
+   * A signal spanning the call is what holds the wait to this: declining
+   * replay settles only the deliveries the caller owns, and a 429 or a 408
+   * is replayed regardless. Measured 2026-09-23: 285-395 ms once connected,
+   * 1181 ms on the first call of a process.
    */
-  judge_likelihood_timeout_ms: z.number().min(1).max(MAX_TIMER_MS).default(10_000),
+  judge_likelihood_timeout_ms: z.number().int().min(1).max(MAX_TIMER_MS).default(10_000),
   /**
    * How much the model may write about one piece of media, in tokens.
    *
